Initial board display + dimensions
We start by making a design decision -how do we model the board that
tiles are save don?

To this end, we'll use a sparse matrix, implemented as a dictionary from
the location (x, y) to the tile itself (temporarily represented by a
character).

That decided, we can display a basic board -- but first, we need the
board's dimensions, so we know what X/Y range to draw in.