Added 'available_spots' function and test
To ensure a user's move is legal, we need to know the list of valid
locations a new tile could go.

available_spots() addresses the first half of this problem: figuring out
all of the potential spots that have an existing tile neighbor but are not
themselves occupied.

Next, we'll need to figure out if our tile actually __fits__ into the
position, but for that we'll need to define the Tile class more cleanly.